Add support for mutation test.
This changeset sets up the tools and the necessary things in the build
pipeline. A script currently added for doing the test but it takes the
existing test cases available but they are not suitable for doing muta
-tion test. Seperate test cases should be written to fully utilize this
feature.

Currently the tool taken for doing mutation test is mull-cxx. It works
with only clang, therefore it's been put into the Linux-clang build
phase.